move accessors to query-accessors
